+++ packages/trpc/src/routers/index.ts
@@ -2,10 +2,12 @@
 import { createTRPCRouter } from "../index";
 import { activitiesRouter } from "./activities";
 import { activityStreamsRouter } from "./activity_streams";
+import { analyticsRouter } from "./analytics";
 import { authRouter } from "./auth";
 import { plannedActivitiesRouter } from "./planned_activities";
 import { profilesRouter } from "./profiles";
 import { storageRouter } from "./storage";
+import { syncRouter } from "./sync";
 
 export const appRouter = createTRPCRouter({
   auth: authRouter,
@@ -14,6 +16,8 @@ export const appRouter = createTRPCRouter({
   activityStreams: activityStreamsRouter,
   plannedActivities: plannedActivitiesRouter,
   storage: storageRouter,
+  sync: syncRouter,
+  analytics: analyticsRouter,
 });
 
 export type AppRouter = typeof appRouter;